working zoom system (still some bugs)

+++ include/GameConstants.hpp
@@ -1,4 +1,4 @@
 #pragma once
 
-static constexpr int TILE_SIZE_PIXELS_UNSCALED = 16;
-static constexpr int CHUNK_TILE_SIZE = 8;
+static constexpr float TILE_SIZE_PIXELS_UNSCALED = 16;
+static constexpr float CHUNK_TILE_SIZE = 8;
